+++ source/core/Triangle.h
@@ -6,7 +6,7 @@ private:
 	bool linearControl;
 	int linearReload;
 	int linearCount;
-	int* location;
+	double* location;
 	bool enable,block;
 	int timer = 0,delayedChange=0;
 public:
@@ -18,7 +18,7 @@ public:
 		15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0,
 			0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15};
 
-	Triangle(int* loc) {
+	Triangle(double* loc) {
 		location = loc;
 	}
 	void registerWrite(int index, uint8_t b, int clock) {
